aem65-an-abbottstore-apps/core/src/main/java/com/abbott/aem/an/abbottstore/models/AdvertisementSectionModel.java
package com.abbott.aem.an.abbottstore.models;

import org.apache.sling.api.resource.Resource;
import org.apache.sling.models.annotations.DefaultInjectionStrategy;
import org.apache.sling.models.annotations.Model;
import org.apache.sling.models.annotations.injectorspecific.ValueMapValue;


/**
 * The Class AdvertisementSectionModel.
 */
@Model(adaptables = { Resource.class }, defaultInjectionStrategy = DefaultInjectionStrategy.OPTIONAL)
public class AdvertisementSectionModel {
	
	/** The title. */
	@ValueMapValue
	private String title;
	
	/** The text. */
	@ValueMapValue
	private String text;

	/**
	 * Gets the title.
	 *
	 * @return the title
	 */
	public String getTitle() {
		return title;
	}

	/**
	 * Gets the text.
	 *
	 * @return the text
	 */
	public String getText() {
		return text;
	}
}
